chore: add JWT authentication to every route
